chore: extend hang watchdog beyond tests/system/
Relocate the exit-hang watchdog (force_exit_if_stalled / pytest_sessionfinish)
from tests/system/conftest.py to the root tests/conftest.py so it protects
every suite, not just system tests. A leaked non-daemon thread in unit,
integration, or e2e tests previously had zero diagnostic coverage —
confirmed directly by observing tests/integration/ finish in 255s while the
pytest process never exited.

Fixes #1707
